Optimize performnce: Create krystal-common module and move data and datatype classes to this new module so that vajram and krystex can reuse them
