+++ common/src/main/java/module-info.java
@@ -1,0 +1,7 @@
+module flipkart.krystal.common {
+  requires com.google.common;
+  requires lombok;
+  exports com.flipkart.krystal.data;
+  exports com.flipkart.krystal.datatypes;
+  exports com.flipkart.krystal.utils;
+}
