+++ vajram/build.gradle
@@ -12,6 +12,7 @@ repositories {
 }
 
 dependencies {
+    implementation project(':common')
     implementation 'com.google.guava:guava:31.1-jre'
     implementation 'org.reflections:reflections:0.10.2'
     implementation 'com.graphql-java:graphql-java:20.0'
