+++ vajram/src/main/java/com/flipkart/krystal/vajram/ExecutionContextMap.java
@@ -1,7 +1,8 @@
 package com.flipkart.krystal.vajram;
 
-import static com.flipkart.krystal.vajram.ValueOrError.empty;
+import static com.flipkart.krystal.data.ValueOrError.empty;
 
+import com.flipkart.krystal.data.ValueOrError;
 import com.flipkart.krystal.vajram.inputs.InputValues;
 import com.google.common.collect.ImmutableMap;
 import java.util.Optional;
