+++ vajram/src/main/java/com/flipkart/krystal/vajram/VajramInputs.java
@@ -1,5 +1,6 @@
 package com.flipkart.krystal.vajram;
 
+import com.flipkart.krystal.data.ValueOrError;
 import com.google.common.collect.ImmutableMap;
 import java.util.Optional;
