+++ vajram/src/main/java/com/flipkart/krystal/vajram/inputs/InputValues.java
@@ -1,8 +1,8 @@
 package com.flipkart.krystal.vajram.inputs;
 
-import static com.flipkart.krystal.vajram.ValueOrError.empty;
+import static com.flipkart.krystal.data.ValueOrError.empty;
 
-import com.flipkart.krystal.vajram.ValueOrError;
+import com.flipkart.krystal.data.ValueOrError;
 import com.google.common.collect.ImmutableMap;
 import java.util.Optional;
 
@@ -17,7 +17,7 @@ public record InputValues(ImmutableMap<String, ValueOrError<?>> values) {
     return (ValueOrError<T>) values().getOrDefault(inputName, empty());
   }
 
-  public  <T> Optional<T> getOpt(String inputName) {
+  public <T> Optional<T> getOpt(String inputName) {
     return this.<T>getValue(inputName).value();
   }
