+++ vajram/src/main/java/module-info.java
@@ -1,15 +1,14 @@
 module flipkart.krystal.vajram {
   exports com.flipkart.krystal.vajram.inputs;
   exports com.flipkart.krystal.vajram;
-  exports com.flipkart.krystal.datatypes;
   exports com.flipkart.krystal.vajram.das;
   exports com.flipkart.krystal.vajram.modulation;
   exports com.flipkart.krystal.vajram.exec;
-  exports com.flipkart.krystal.utils;
 
   requires com.google.common;
   requires static lombok;
   requires org.reflections;
   requires org.checkerframework.checker.qual;
   requires com.google.errorprone.annotations;
+  requires flipkart.krystal.common;
 }
