+++ vajram/vajram-codegen/build.gradle
@@ -26,6 +26,7 @@ gradlePlugin {
 
 
 dependencies {
+    implementation project(':common')
     implementation project(':vajram')
     implementation 'com.google.guava:guava:31.1-jre'
     implementation 'com.squareup:javapoet:1.13.0'
