+++ vajram/vajram-krystex/build.gradle
@@ -25,6 +25,7 @@ repositories {
 }
 
 dependencies {
+    implementation project(':common')
     implementation project(':vajram')
     implementation project(':krystex')
     implementation 'com.google.guava:guava:31.1-jre'
