+++ vajram/vajram-samples/build.gradle
@@ -20,6 +20,7 @@ repositories {
 }
 
 dependencies {
+    implementation project(':common')
     implementation project(':vajram')
     implementation 'com.google.guava:guava:31.1-jre'
