+++ vajram/vajram-samples/src/main/java/module-info.java
@@ -1,5 +1,6 @@
 module flipkart.krystal.vajram.vajram_samples {
   requires flipkart.krystal.vajram;
+  requires flipkart.krystal.common;
   requires com.google.common;
   requires static lombok;
   requires java.logging;
